Updated requirements to hotfixed golem-messages
Removed tests migrated to golem-messages
